Renames meta by version_control
