+++ api/specs/webserver/openapi-projects.yaml
@@ -443,7 +443,7 @@ paths:
         - snapshot
       summary: List Snapshots
       description: Lists all snapshots taken from a given project
-      operationId: simcore_service_webserver.meta_api_handlers_snapshots.list_project_snapshots_handler
+      operationId: simcore_service_webserver.version_control_handlers_snapshots.list_project_snapshots_handler
       parameters:
         - required: true
           schema:
@@ -495,7 +495,7 @@ paths:
         - snapshot
       summary: Create Snapshot
       description: Takes a snapshot of the project at this time
-      operationId: simcore_service_webserver.meta_api_handlers_snapshots.create_project_snapshot_handler
+      operationId: simcore_service_webserver.version_control_handlers_snapshots.create_project_snapshot_handler
       parameters:
         - required: true
           schema:
@@ -531,7 +531,7 @@ paths:
         - snapshot
       summary: Get Snapshot
       description: Gets commit info for a given snapshot
-      operationId: simcore_service_webserver.meta_api_handlers_snapshots.get_project_snapshot_handler
+      operationId: simcore_service_webserver.version_control_handlers_snapshots.get_project_snapshot_handler
       parameters:
         - required: true
           schema:
@@ -566,7 +566,7 @@ paths:
         - snapshot
       summary: Delete Snapshot
       description: Deletes both the commit and the project itself
-      operationId: simcore_service_webserver.meta_api_handlers_snapshots.delete_project_snapshot_handler
+      operationId: simcore_service_webserver.version_control_handlers_snapshots.delete_project_snapshot_handler
       parameters:
         - required: true
           schema:
@@ -597,7 +597,7 @@ paths:
         - snapshot
       summary: Update Snapshot
       description: Updates label/name of a snapshot
-      operationId: simcore_service_webserver.meta_api_handlers_snapshots.patch_project_snapshot_handler
+      operationId: simcore_service_webserver.version_control_handlers_snapshots.patch_project_snapshot_handler
       parameters:
         - required: true
           schema:
