+++ services/web/server/src/simcore_service_webserver/api/v0/openapi.yaml
@@ -13258,7 +13258,7 @@ paths:
         - snapshot
       summary: List Snapshots
       description: Lists all snapshots taken from a given project
-      operationId: simcore_service_webserver.meta_api_handlers_snapshots.list_project_snapshots_handler
+      operationId: simcore_service_webserver.version_control_handlers_snapshots.list_project_snapshots_handler
       parameters:
         - required: true
           schema:
@@ -13384,7 +13384,7 @@ paths:
         - snapshot
       summary: Create Snapshot
       description: Takes a snapshot of the project at this time
-      operationId: simcore_service_webserver.meta_api_handlers_snapshots.create_project_snapshot_handler
+      operationId: simcore_service_webserver.version_control_handlers_snapshots.create_project_snapshot_handler
       parameters:
         - required: true
           schema:
@@ -13493,7 +13493,7 @@ paths:
         - snapshot
       summary: Get Snapshot
       description: Gets commit info for a given snapshot
-      operationId: simcore_service_webserver.meta_api_handlers_snapshots.get_project_snapshot_handler
+      operationId: simcore_service_webserver.version_control_handlers_snapshots.get_project_snapshot_handler
       parameters:
         - required: true
           schema:
@@ -13602,7 +13602,7 @@ paths:
         - snapshot
       summary: Delete Snapshot
       description: Deletes both the commit and the project itself
-      operationId: simcore_service_webserver.meta_api_handlers_snapshots.delete_project_snapshot_handler
+      operationId: simcore_service_webserver.version_control_handlers_snapshots.delete_project_snapshot_handler
       parameters:
         - required: true
           schema:
@@ -13657,7 +13657,7 @@ paths:
         - snapshot
       summary: Update Snapshot
       description: Updates label/name of a snapshot
-      operationId: simcore_service_webserver.meta_api_handlers_snapshots.patch_project_snapshot_handler
+      operationId: simcore_service_webserver.version_control_handlers_snapshots.patch_project_snapshot_handler
       parameters:
         - required: true
           schema:
@@ -13728,7 +13728,7 @@ paths:
         - repository
       summary: List Repos
       description: List info about versioned projects
-      operationId: simcore_service_webserver.meta_api_handlers_repos._list_repos_handler
+      operationId: simcore_service_webserver.version_control_handlers._list_repos_handler
       parameters:
         - description: index to the first item to return (pagination)
           required: false
@@ -13885,7 +13885,7 @@ paths:
         - repository
       summary: List Checkpoints
       description: Lists commits&tags tree of the project
-      operationId: simcore_service_webserver.meta_api_handlers_repos._list_checkpoints_handler
+      operationId: simcore_service_webserver.version_control_handlers._list_checkpoints_handler
       parameters:
         - description: Project unique identifier
           required: true
@@ -14071,7 +14071,7 @@ paths:
       tags:
         - repository
       summary: Create Checkpoint
-      operationId: simcore_service_webserver.meta_api_handlers_repos._create_checkpoint_handler
+      operationId: simcore_service_webserver.version_control_handlers._create_checkpoint_handler
       parameters:
         - description: Project unique identifier
           required: true
@@ -14195,7 +14195,7 @@ paths:
         - repository
       summary: Get Checkpoint
       description: Set ref_id=HEAD to return current commit
-      operationId: simcore_service_webserver.meta_api_handlers_repos._get_checkpoint_handler
+      operationId: simcore_service_webserver.version_control_handlers._get_checkpoint_handler
       parameters:
         - description: 'A repository ref (commit, tag or branch)'
           required: true
@@ -14312,7 +14312,7 @@ paths:
       tags:
         - repository
       summary: Update Checkpoint Annotations
-      operationId: simcore_service_webserver.meta_api_handlers_repos._update_checkpoint_annotations_handler
+      operationId: simcore_service_webserver.version_control_handlers._update_checkpoint_annotations_handler
       parameters:
         - description: 'A repository ref (commit, tag or branch)'
           required: true
@@ -14447,7 +14447,7 @@ paths:
       description: |-
         Affect current working copy of the project, i.e. get_project will now return
         the check out
-      operationId: simcore_service_webserver.meta_api_handlers_repos._checkout_handler
+      operationId: simcore_service_webserver.version_control_handlers._checkout_handler
       parameters:
         - description: 'A repository ref (commit, tag or branch)'
           required: true
@@ -14566,7 +14566,7 @@ paths:
         - repository
       summary: View Project Workbench
       description: Returns a view of the workbench for a given project's version
-      operationId: simcore_service_webserver.meta_api_handlers_repos._view_project_workbench_handler
+      operationId: simcore_service_webserver.version_control_handlers._view_project_workbench_handler
       parameters:
         - description: 'A repository ref (commit, tag or branch)'
           required: true
