+++ services/web/server/src/simcore_service_webserver/application_config.py
@@ -94,7 +94,7 @@ def create_schema() -> T.Dict:
             addon_section("studies_access", optional=True): minimal_addon_schema(),
             addon_section("studies_dispatcher", optional=True): minimal_addon_schema(),
             addon_section("exporter", optional=True): minimal_addon_schema(),
-            addon_section("meta", optional=True): minimal_addon_schema(),
+            addon_section("version_control", optional=True): minimal_addon_schema(),
             addon_section("clusters", optional=True): minimal_addon_schema(),
         }
     )
