+++ services/web/server/src/simcore_service_webserver/version_control_core_snapshots.py
@@ -6,15 +6,21 @@
 
 """
 
-
+import warnings
 from datetime import datetime
 from typing import Any, Dict, Optional, Tuple
 from uuid import UUID
 
 from models_library.projects import Project
 
-from .meta_models_snapshots import Snapshot
 from .projects import projects_utils
+from .version_control_models_snapshots import Snapshot
+
+warnings.warn(
+    "version_control_*_snapshots.py modules are the first generation of vc."
+    "It is just temporarily kept it functional until it gets fully replaced",
+    DeprecationWarning,
+)
 
 ProjectDict = Dict[str, Any]
