+++ services/web/server/tests/unit/with_dbs/10/meta/test_version_control.py
@@ -7,15 +7,17 @@ import itertools
 import pytest
 from openapi_core.schema.specs.models import Spec as OpenApiSpecs
 from simcore_service_webserver._meta import api_vtag as vtag
-from simcore_service_webserver.meta import (
-    meta_api_handlers_repos,
-    meta_api_handlers_snapshots,
+from simcore_service_webserver.version_control import (
+    version_control_handlers,
+    version_control_handlers_snapshots,
 )
 
 
 @pytest.mark.parametrize(
     "route",
-    itertools.chain(meta_api_handlers_repos.routes, meta_api_handlers_snapshots.routes),
+    itertools.chain(
+        version_control_handlers.routes, version_control_handlers_snapshots.routes
+    ),
     ids=lambda r: f"{r.method.upper()} {r.path}",
 )
 def test_route_against_openapi_specs(route, openapi_specs: OpenApiSpecs):
