+++ services/web/server/tests/unit/with_dbs/10/meta/test_version_control_handlers_snapshots.py
@@ -10,7 +10,10 @@ from aiohttp import web
 from models_library.projects import Project
 from pytest_simcore.helpers.utils_assert import assert_status
 from simcore_service_webserver._meta import api_vtag as vtag
-from simcore_service_webserver.meta_models_snapshots import SnapshotItem, SnapshotPatch
+from simcore_service_webserver.version_control_models_snapshots import (
+    SnapshotItem,
+    SnapshotPatch,
+)
 
 ProjectDict = Dict[str, Any]
